[refactor] 패키지 구조 변경
헥사고날 아키텍처 스타일에 맞게 구조 변경
팀 내 논의에 따라 response -> results로 용어 통일
