+++ src/main/java/com/prgrms/ohouse/domain/commerce/model/review/Review.java
@@ -81,4 +81,3 @@ public class Review implements ImageAttachable {
 		return reviewImage;
 	}
 }
-
