+++ src/main/java/com/prgrms/ohouse/domain/common/file/FileRepository.java
@@ -1,8 +1,5 @@
 package com.prgrms.ohouse.domain.common.file;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.prgrms.ohouse.domain.common.file.StoredFile;
-
-public interface FileRepository extends JpaRepository<StoredFile, Long> {
+public interface FileRepository {
+	StoredFile save(StoredFile storedFile);
 }
