--- src/main/java/com/prgrms/ohouse/domain/community/application/dummy.java
@@ -1,4 +1,0 @@
-package com.prgrms.ohouse.domain.community.application;
-
-public class dummy {
-}
