+++ src/main/java/com/prgrms/ohouse/domain/community/model/housewarming/Family.java
@@ -1,4 +1,4 @@
-package com.prgrms.ohouse.domain.community.model.post.hwpost;
+package com.prgrms.ohouse.domain.community.model.housewarming;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
@@ -18,7 +18,7 @@ public class Family {
 	@Column(name = "family_count")
 	private Integer memberCount;
 
-	enum Type {
+	public enum Type {
 		SINGLE, MARRIED, CHILDREN, PARENTS, ROOMMATES
 	}
 }
