+++ src/main/java/com/prgrms/ohouse/domain/community/model/housewarming/HousewarmingPost.java
@@ -1,4 +1,4 @@
-package com.prgrms.ohouse.domain.community.model.post.hwpost;
+package com.prgrms.ohouse.domain.community.model.housewarming;
 
 import java.util.ArrayList;
 import java.util.List;
