+++ src/main/java/com/prgrms/ohouse/domain/community/model/housewarming/WorkerType.java
@@ -1,0 +1,5 @@
+package com.prgrms.ohouse.domain.community.model.housewarming;
+
+public enum WorkerType {
+	SELF, HALF_SELF, PRO
+}
