--- src/main/java/com/prgrms/ohouse/domain/community/model/post/hwpost/WorkerType.java
@@ -1,5 +1,0 @@
-package com.prgrms.ohouse.domain.community.model.post.hwpost;
-
-public enum WorkerType {
-	SELF, HALF_SELF, PRO
-}
