+++ src/main/java/com/prgrms/ohouse/domain/community/model/question/QuestionPostImage.java
@@ -1,4 +1,4 @@
-package com.prgrms.ohouse.domain.community.model.post.question;
+package com.prgrms.ohouse.domain.community.model.question;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
