+++ src/main/java/com/prgrms/ohouse/domain/community/model/question/QuestionPostRepository.java
@@ -1,0 +1,7 @@
+package com.prgrms.ohouse.domain.community.model.question;
+
+public interface QuestionPostRepository {
+	QuestionPost save(QuestionPost questionPost);
+
+	void deleteById(Long id);
+}
